refactor(shared): consolidate inference time helpers (#30125)
