+++ packages/shared/src/lifeops-normalize/index.ts
@@ -9,4 +9,5 @@
 
 export * from "./service-error.js";
 export * from "./service-normalize.js";
+export * from "./time-util.js";
 export * from "./time-zone.js";
